twinsApi/requirements.txt
fastapi
uvicorn
python-dotenv
PyJWT
bcrypt